MonadResult

Description

This module mainly export the MonadResult type class, which represent the output of some algorithm that may be interpreted as a failure or a success.

The idea is to generalize some common usage of sum types such as Maybe and Either, and make more general some algorithm that handle result transfer without doing too much result inquiry (e.g., a parser).

Documentation

class Monad m => MonadResult m where

The MonadResult type class. It involves m being a Monad mainly because instances are typically used in a monadic setting.

Minimal complete definition

isFailure, extract, extractM

Methods

isFailure :: m a -> Bool

Test if the monad is in a failure(-type) state.

isSuccess :: m a -> Bool

Test if the monad is in a success(-type) state. By default, it simply negate the isFailure function. It may be useful to override that for "multi-valued logic" kind of types.

extract :: m a -> a

Extract the value inside the monad. Caller of this method should enforce that the monad is in a success(-type) state (using isSuccess). Conversely, this function is only called on a success monad.

extractM :: m a -> Maybe a

Extract the value inside the monad _if possible_, or return Nothing if impossible. If the monad is in a success(-type) state, must return a Just. Otherwise, must return a Nothing.