Parsing

Description

This library implements a simple-yet-versatile form of parser combinators, represented basically as state monads. Parsers are alternative monads, with <|> being the choice (alternative in a grammar) and >>= being the sequencing of parsers.

The module also provides alternative functions and operators, for convenience. For instance, %> and %:> exploit parsing results that are MonadPlus to parse and accumulate results (which is very common when having repetitions, for instance).

The library has been made fairly generic, in particular in term of how the parsing result is made. A "default" implementation uses the Maybe monad, which is good enough for simple parsing.

Documentation

newtype ParserT m s r

The main type for the parser, or parser "transformer". The first type parameter (m) is the monad wrapping the parsing result. It should be some kind of alternative monad or MonadFail in theory. This is not mandatory but will greatly impact the lib's usability.

The second type parameter (s) is the incoming stream. Once again, this could really be anything, although a number of convenient functions work on _stream-like_ inputs (see Stream typeclass).

Last, the third type parameter (r) is the type of the outgoing result. This is the type that varies the most and for which ParserT is a well-behave monad.

Constructors

ParserT

Fields

runParserT :: s -> m (s, r)

Unwrap the parser, i.e. runs it

execParserT :: Functor m => ParserT m s r -> s -> m r

Execute the parser, i.e. run it and keep only the result (if available). Note that no verification is done as if there is still stream to be parsed. A good grammar should explicitly have a starting rule with a end or similar terminal, indicating the end of stream.

type Parser = ParserT Maybe

Type synonym for parsers which wrapping monad is Maybe.

runParser :: Parser s r -> s -> Maybe (s, r)

Run a parser with Maybe as its monad

parserTOf :: (s -> m (s, r)) -> ParserT m s r

Create a parser from a function.

parserOf :: (s -> Maybe (s, r)) -> Parser s r

Same as parserTOf but specialized for parser with Maybe

class Stream s where

A class representing a certain type of polymorphic data structure for which it is possible to get an element + the rest of the structure (presumably without the element).

This class effectively factorises anything that looks like a sequence (or a list). It allows for a bit of flexibility when using parsers (rather than using plain old lists).

Methods

uncons :: s a -> Maybe (a, s a)

Takes the stream and try to extract an element and the remainder of the stream. If not possible, return Nothing.

This function returning Nothing implies that the stream is finished (end of string, end of file, etc.) or at least that no other character is available.

done :: Stream s => s a -> Bool

Tests if the stream is done, i.e. it is empty/no other symbol is available

failIf :: Alternative m => (s -> Bool) -> ParserT m s ()

Parser that fails if the predicate is true on the given stream, and succeeds while consuming nothing otherwise.

zero :: (Monad m, MonadPlus f) => ParserT m s (f a)

Parser that always succeeds, consumes nothing and returns mzero in the provided MonadPlus instance. This is especially useful for repetition, i.e. : > -- A -> / (empty word) > -- A -> x A > parseA = (parseChar x >>= x -> parseA >>= xs -> x mplus xs) | zero

end :: (Alternative m, Stream s) => ParserT m (s a) ()

Parser that succeeds iff the stream has reach its end (and does not consume anything, subsequently).

jump :: Monad m => s -> ParserT m s ()

Jump in the incoming stream, i.e. replace the current stream by another one. This is intended for redirections during parsing, and especially resetting the incoming stream (when performing simple lookahead).

This parser always succeeds, consume nothing and returns ()

drop1 :: (Stream s, Alternative m) => ParserT m (s a) ()

Drop a symbol of the stream. This parser fails if the stream is done.

keep1 :: (Stream s, Alternative m, Applicative f) => ParserT m (s a) (f a)

Consume one symbol of the stream and return it as result (as a MonadPlus). Fails if the stream is done.

This is intended for "transponder"-type parsers, which rewrites the input stream on their output (after transformation). The result is a MonadPlus to allow to chain the keep1. The result is usually a list as well in that case.

notEmpty :: (Alternative m, Stream s) => ParserT m (s a) ()

Parser that consumes nothing and fails if the stream is done/empty.

parseAndBack :: Monad m => ParserT m s r -> ParserT m s r

Parser combinator that runs the given parser and then sets back the stream where it was before parsing.

For instance, if the result of parsing `onexyz...` is `1` with remaining stream being `xyz...`, then using parseAndBack on this parser will yield the result `1` (result of the parser) but `onexyz...` as the remaining stream.

(%>) :: (Monad m, MonadPlus f) => ParserT m s (f a) -> ParserT m s (f a) -> ParserT m s (f a) infixr 3

Convenient combinator for parsers which result is a MonadPlus, that chains the parsers and mplus their results.

(%:>) :: (Monad m, MonadPlus f) => ParserT m s a -> ParserT m s (f a) -> ParserT m s (f a) infixr 3

Specific version of %> where the first argument is not wrapped in the MonadPlus (convenient for injecting an element in a list, typically).

parseP :: (Alternative m, Stream s) => (a -> Bool) -> ParserT m (s a) a

Parser that succeeds if the head of the stream satisfies the given predicate. If it does not or if the stream is done, it fails. This returns the consumed symbol.

parseP_ :: (Alternative m, Stream s) => (a -> Bool) -> ParserT m (s a) ()

Same as parserP but ignores the parsed symbol (and returns ()).

parseChar :: (Alternative m, Stream s, Eq a) => a -> ParserT m (s a) a

Parser that succeeds if the head of the stream is the given symbol, and fails otherwise or if the stream is done.

parseChar_ :: (Alternative m, Stream s, Eq a) => a -> ParserT m (s a) ()

Same as parseChar but drops the symbol and returns ().

parseSeq :: (Monad m, Alternative m, Eq a, Foldable t, Stream s, MonadPlus f) => t a -> ParserT m (s a) (f a)

Parser that succeeds if the first elements of the stream correspond to the given sequence of symbols. If the stream is empty (or not long enough) or if the sequence is not right, it fails.

The returned result correspond to each character of the sequence that was parsed successfully, wrapped in the MonadPlus result.

parseSeq_ :: (Monad m, Alternative m, Eq a, Foldable t, Stream s) => t a -> ParserT m (s a) ()

Same as parseSeq but drops the result and only return ().

parseSpace :: (Alternative m, Stream s) => ParserT m (s Char) ()

Parse exactly one blank space (and drop it).

parseSpaces :: (Monad m, Alternative m, Stream s) => ParserT m (s Char) ()

Parse a sequence of blank spaces (at least one) and drop the result.

parseDigit :: (Alternative m, Stream s) => ParserT m (s Char) Int

Parse one digit and return the corresponding Int.

parseNumber :: (Monad m, Alternative m, Stream s) => ParserT m (s Char) Int

Parse a full number (and return its value). A number in this parser is simply a sequence of digits with no further constraints. I do not think it works on hexadecimal numbers (?).

parseShow :: (Monad m, Alternative m, Show x, Stream s) => x -> ParserT m (s Char) x

Parse a type based on its textual representation given by its Show typeclass intsance.