Parsing.Result

Description

This module defines the Result type, specialized for parsing. This type differentiates between success, nothing parsed and actual failure, allowing for much ritcher behaviours.

The type is also embedded in a specific parser type synonym with context'd streams (see Ctx), which provide useful metadatas for errors raised.

Documentation

data Error

Record type holding an error message with metadata

Constructors

Error

Fields

error_source :: String

Source of the error (usually a file or user input)

error_line :: Int

Line number of the error

error_column :: Int

Column number of the error (the error is usually _before_ due to how parsers work

error_message :: String

Error message

error_cause :: Maybe Error

Possible cause for the error (allowing for error hierarchy, similar in principle to Java exceptions)

mkError :: String -> Int -> Int -> String -> Maybe Error -> Error

Make an error using each components of the record (this is in case the order of the fields changes).

mkError' :: String -> Int -> Int -> String -> Error

Make an error without a cause.

data Result a

The result type. The result of an algorithm (parser) may be either a success (with associated computed element), a failure (with associated error), or simply a IOU/nothing was parsed result.

Constructors

Success a

NoParse

Failure Error

type ParserR = ParserT Result

Type synonym for parsers that use Result as monad.

type ParserCtx s a r = ParserCtxT Result s a r

Type synonym for parsers that contex' streams and use Result as monad.

runParserCtx :: Stream s => ParserCtx s a r -> (a -> Bool) -> String -> (Int, Int) -> s a -> Result (Ctx s a, r)

Specialized version of runParserCtxT for the Result monad.

execParserCtx :: Stream s => ParserCtx s a r -> (a -> Bool) -> String -> (Int, Int) -> s a -> Result r

Specialized version of execParserCtxT for the Result monad.

runParserCtxC :: Stream s => ParserCtx s Char r -> String -> (Int, Int) -> s Char -> Result (Ctx s Char, r)

Specialized version of runParserCtx for character streams (usual newline symbol used).

execParserCtxC :: Stream s => ParserCtx s Char r -> String -> (Int, Int) -> s Char -> Result r

Specialized version of execParserCtx for character streams (usual newline symbol used)

raise :: String -> ParserCtx s a r

Create a parser that raises a failure. This is typically used as the last element in a <|> chain to notify that the parsing did not work while it should have.

orDie :: ParserCtx s a r -> String -> ParserCtx s a r infix 2

Run a parser and raise an error if it does not succeed. This is to be used in `do` notations

do
  r <- parseNumber `orDie` "Expecting a number here"
  parseChar '+'
  i <- parseNumber `orDie` "Expecting imaginary part"
  parseChar 'i'
  return $ Complexe r i

To be precise, this function creates a new error if the result of the parser is NoParse, and if the result is Failure, then it creates a new error which cause is the error of that failure.