Parsing.Stream

Documentation

class Stream s where

A class representing a certain type of polymorphic data structure for which it is possible to get an element + the rest of the structure (presumably without the element).

This class effectively factorises anything that looks like a sequence (or a list). It allows for a bit of flexibility when using parsers (rather than using plain old lists).

Methods

uncons :: s a -> Maybe (a, s a)

Takes the stream and try to extract an element and the remainder of the stream. If not possible, return Nothing.

This function returning Nothing implies that the stream is finished (end of string, end of file, etc.) or at least that no other character is available.

done :: Stream s => s a -> Bool

Tests if the stream is done, i.e. it is empty/no other symbol is available

data Ctx s a

A Stream wrapper that embeds contextual informations (i.e. source, line number and column number). The idea is that uppon uncons'ing, the values move by themselves, allowing parsers to access a context, mainly for error reporting.

Constructors

Ctx

Fields

_stream :: s a

Wrapped stream

_newline :: a -> Bool

Predicate establishing is a symbol count as a newline

ctx_source :: String

Source of the stream (e.g. a file, user input, etc.)

ctx_line :: Int

Line number for the current character

ctx_column :: Int

Column number for the current character

withContext :: Stream s => (a -> Bool) -> String -> (Int, Int) -> s a -> Ctx s a

Wrap a stream into a Ctx, with given starting coordinates and newline predicate.

withContextC :: Stream s => String -> (Int, Int) -> s Char -> Ctx s Char

Wrap a Char stream into a Ctx, with given starting coordinates. Newline predicate is simply `== 'n'`.

contextStart :: (Int, Int)

Usual starting coordinates for a context (line 1, column 1).